Update instructions.md
The default prompt was failing to "keep" the name in the structure was already proper ie Last, First. It was swapping the names all times.

The given new prompt:
```
A user has input their first and last name into a form. We don't know in which order their first name and last name are, but we need it to be in this format '<Last name>, <First name>'. 

Please convert the following name in the expected format:

doe john
```
Performs as expected.